+++ src/main/java/com/cburch/logisim/gui/log/ComponentSelector.java
@@ -305,9 +305,11 @@ public class ComponentSelector extends JTable {
       return option.toString();
     }
   }
-
+  
+  //Suppress serial warning because TreeNodeRenderer is a UI component and 
+  //is not intended to be serialized or persisted across different JVM versions.
+  @SuppressWarnings("serial")
   private class TreeNodeRenderer extends DefaultTableCellRenderer implements Icon {
-
     private TreeNode<?> node;
 
     @Override
@@ -315,7 +317,7 @@ public class ComponentSelector extends JTable {
       if (value instanceof CircuitNode) isSelected = false;
       final java.awt.Component ret = super.getTableCellRendererComponent(table, value, isSelected, hasFocus, row, column);
       if (ret instanceof JLabel && value instanceof TreeNode) {
-        node = (TreeNode) value;
+        node = (TreeNode<?>) value;
         ((JLabel) ret).setIcon(this);
       }
       return ret;